refactor train.py to reduce LoCs.
