+++ helpers/training/adapter.py
@@ -1,0 +1,47 @@
+from peft import LoraConfig
+
+def determine_adapter_target_modules(args, unet, transformer):
+    if unet is not None:
+        return ["to_k", "to_q", "to_v", "to_out.0"]
+    elif transformer is not None:
+        target_modules = ["to_k", "to_q", "to_v", "to_out.0"]
+        
+        if args.flux and args.flux_lora_target == "all":
+            # target_modules = mmdit layers here
+            target_modules = [
+                "to_k",
+                "to_q",
+                "to_v",
+                "add_k_proj",
+                "add_q_proj",
+                "add_v_proj",
+                "to_out.0",
+                "to_add_out.0",
+            ]
+        elif args.flux_lora_target == "context":
+            # i think these are the text input layers.
+            target_modules = [
+                "add_k_proj",
+                "add_q_proj",
+                "add_v_proj",
+                "to_add_out.0",
+            ]
+        elif args.flux_lora_target == "all+ffs":
+            target_modules = [
+                "to_k",
+                "to_q",
+                "to_v",
+                "add_k_proj",
+                "add_q_proj",
+                "add_v_proj",
+                "to_out.0",
+                "to_add_out.0",
+                "ff.0",
+                "ff.2",
+                "ff_context.0",
+                "ff_context.2",
+                "proj_mlp",
+                "proj_out",
+            ]
+
+        return target_modules
